# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Location {
  /** Location ID e.g. A1-01-02 */
  id: string;
  description?: string | null;
  createdAt: string;
}

export interface CreateLocationRequest {
  id: string;
  description?: string | null;
}

export interface StockEntry {
  locationId: string;
  productId: number;
  quantity: number;
  productName: string;
  productCategory: string;
  bufferStock: number;
}

export interface LocationWithStock {
  id: string;
  description?: string | null;
  createdAt: string;
  stock: StockEntry[];
}

export interface Product {
  id: number;
  name: string;
  category: string;
  /** Minimum required quantity */
  bufferStock: number;
  alertEmail?: string | null;
  createdAt: string;
}

export interface ProductWithStock {
  id: number;
  name: string;
  category: string;
  bufferStock: number;
  alertEmail?: string | null;
  createdAt: string;
  /** Total quantity across all locations */
  totalStock: number;
  isLowStock: boolean;
}

export interface CreateProductRequest {
  name: string;
  category: string;
  bufferStock: number;
  alertEmail?: string | null;
}

export interface UpdateProductRequest {
  name?: string;
  category?: string;
  bufferStock?: number;
  alertEmail?: string | null;
}

export interface UpdateStockRequest {
  /** New absolute quantity */
  quantity?: number;
  /** Relative change (positive to add, negative to remove) */
  delta?: number;
  changedBy?: string | null;
}

export interface StockUpdateResult {
  locationId: string;
  productId: number;
  quantity: number;
  totalStock: number;
  isLowStock: boolean;
  alertSent: boolean;
}

export interface HistoryEntry {
  id: number;
  locationId: string;
  productId: number;
  productName: string;
  previousQuantity: number;
  newQuantity: number;
  delta: number;
  changedBy?: string | null;
  changedAt: string;
}

export interface CategoryCount {
  category: string;
  productCount: number;
  totalStock: number;
}

export interface DashboardSummary {
  totalLocations: number;
  totalProducts: number;
  lowStockProducts: ProductWithStock[];
  recentActivity: HistoryEntry[];
  categoryBreakdown: CategoryCount[];
}

export type ListHistoryParams = {
  productId?: number;
  locationId?: string;
  limit?: number;
};