# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  category: string;
  bufferStock: number;
  alertEmail?: string | null;
}