# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from "./categoryCount";
import type { HistoryEntry } from "./historyEntry";
import type { ProductWithStock } from "./productWithStock";

export interface DashboardSummary {
  totalLocations: number;
  totalProducts: number;
  lowStockProducts: ProductWithStock[];
  recentActivity: HistoryEntry[];
  categoryBreakdown: CategoryCount[];
}