# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  locationId: string;
  productId: number;
  productName: string;
  previousQuantity: number;
  newQuantity: number;
  delta: number;
  changedBy?: string | null;
  changedAt: Date;
}