# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  /** Location ID e.g. A1-01-02 */
  id: string;
  description?: string | null;
  createdAt: Date;
}