# lib/api-zod/src/generated/types/stockEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface StockEntry {
  locationId: string;
  productId: number;
  quantity: number;
  productName: string;
  productCategory: string;
  bufferStock: number;
}