# lib/api-zod/src/generated/types/stockUpdateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface StockUpdateResult {
  locationId: string;
  productId: number;
  quantity: number;
  totalStock: number;
  isLowStock: boolean;
  alertSent: boolean;
}