# lib/api-zod/src/generated/types/updateStockRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Warehouse Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStockRequest {
  /** New absolute quantity */
  quantity?: number;
  /** Relative change (positive to add, negative to remove) */
  delta?: number;
  changedBy?: string | null;
}